Huuuge...

+++ src/core/input.rs
@@ -23,14 +23,15 @@ fn update_mouse_state_system(
     mut mouse_motion_events: EventReader<CursorMoved>,
 ) {
     // Only one window and only one camera
-    let window = windows.single();
-    let camera = cameras.single();
+    if let Some(window) = windows.iter().next() {
+        let camera = cameras.single();
 
-    if let Some(e) = mouse_motion_events.read().last() {
-        let x = e.position.x - window.width() / 2.0;
-        let y = window.height() / 2.0 - e.position.y;
-        let translation = camera.1.flat();
-        let scale = camera.1.scale.truncate();
-        mouse_state.position = Some(translation + Vec2::new(x, y) * scale);
+        if let Some(e) = mouse_motion_events.read().last() {
+            let x = e.position.x - window.width() / 2.0;
+            let y = window.height() / 2.0 - e.position.y;
+            let translation = camera.1.flat();
+            let scale = camera.1.scale.truncate();
+            mouse_state.position = Some(translation + Vec2::new(x, y) * scale);
+        }
     }
 }
